+++ assets/js/script.js
@@ -9,7 +9,8 @@ document.addEventListener('DOMContentLoaded', function() {
     if (phoneInput) {
         iti = window.intlTelInput(phoneInput, {
             utilsScript: "https://cdnjs.cloudflare.com/ajax/libs/intl-tel-input/17.0.8/js/utils.js",
-            preferredCountries: ['us', 'gb'],
+            preferredCountries: ['gb', 'us'],
+            initialCountry: 'gb',
             separateDialCode: true,
             autoPlaceholder: "aggressive"
         });